pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "blick"
version = "0.1.0"
description = "Security-focused multi-identity agent framework"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.9",
    "pyyaml>=6.0",
    "rapidfuzz>=3.0",
    "cryptography>=42.0",
    "keyring>=25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "aiosqlite>=0.20",
]

[project.scripts]
blick = "blick.__main__:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "llm: tests that require a running Ollama LLM (use -m llm to run)",
]

[tool.hatch.build.targets.wheel]
packages = ["blick"]
